prototype 3/mobile_app/foodlensback/requirements.txt
streamlit
torch
numpy
Pillow
albumentations
opencv-python-headless
transformers
tqdm
django
djangorestframework
django-cors-headers
torchvision
transformers
sentencepiece

